feat: cleanup-ui-activity-card-coverage-split-001 - Preserve editor chrome and import workflow coverage in focused suites
Split editor chrome and import workflow tests out of the monolithic current
activity card suite into focused files with a shared feature-local harness so
failures localize to the relevant workflow without changing runtime behavior.
