+++ ui/biome.jsonc
@@ -187,6 +187,9 @@
         "src/features/work-outcome/hooks/useWorkOutcomeChart.test.ts",
         "src/features/workflow-activity/components/dashboard-flow-axis-legend.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card-import.test.tsx",
+        "src/features/workflow-activity/components/react-flow-current-activity-card-editor-chrome.test.tsx",
+        "src/features/workflow-activity/components/react-flow-current-activity-card-import-flows.test.tsx",
+        "src/features/workflow-activity/components/react-flow-current-activity-card-component.harness.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card.coverage.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card.test.tsx",
         "src/features/workflow-activity/react-flow-current-activity-card-editor-controllers.test.tsx",
