+++ ui/src/features/workflow-activity/components/react-flow-current-activity-card.test.tsx
@@ -1,16 +1,15 @@
 import "@testing-library/jest-dom/vitest";
+import "./react-flow-current-activity-card-component.mocks.test";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import {
   cleanup,
   fireEvent,
-  render,
   screen,
   waitFor,
   within,
 } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
-import type { ReactElement } from "react";
+import type { CanonicalFactoryDefinition } from "../../../api/current-factory-definition";
 import type {
   DashboardPlaceRef,
   DashboardSnapshot,
@@ -20,8 +19,6 @@ import {
   type ImportFactoryValue,
   SessionFactoryAPIError,
 } from "../../../api/session-factory";
-import { factoryFromDashboardTopology } from "../../../components/dashboard/fixtures";
-import { installDashboardBrowserTestShims } from "../../../components/dashboard/test-browser-shims";
 import {
   resourceOccupancySnapshotForTick,
   semanticWorkflowDashboardSnapshot,
@@ -32,38 +29,24 @@ import {
 } from "../../../components/dashboard/test-fixtures";
 import { DashboardSessionTestProvider } from "../../../testing/dashboard-session-test-provider";
 import {
-  baseFactoryDefinition,
   baseFactoryDefinitionDocument,
-  buildDivergentPlaneDashboardSnapshot,
-  buildMockGraphSavePayload,
-  createMockGraphEditorDraftState,
-  divergentDocumentPlaneFactoryDocument,
-  wireMockEditableFactoryGraph,
   workerDenseFactoryDefinitionDocument,
 } from "../../../testing/graph-editor-harness";
-import { selectLabeledComboboxOption } from "../../../testing/select-test-helpers";
 import { useCurrentFactoryDocument } from "../../current-factory-definition/hooks/useCurrentFactoryDefinition";
-import { useFactoryDocumentSave } from "../../current-factory-definition/hooks/useFactoryDocumentSave";
 import { useFactoryGraphDraftState } from "../../factory-graph-editor/hooks/factory-graph-draft-hook";
-import { useEditableFactoryGraph } from "../../factory-graph-editor/hooks/use-editable-factory-graph";
 import { maintainerRuntimeShapedFactory } from "../../factory-graph-editor/lib/fixtures/maintainer-runtime-shaped-factory.fixture";
 import {
   SYSTEM_TIME_EXPIRY_TRANSITION_ID,
   SYSTEM_TIME_WORK_TYPE_ID,
   systemTimeGraphNodeId,
 } from "../../factory-graph-editor/lib/operations/factory-graph-customer-display";
-import { removeFactoryGraphNode } from "../../factory-graph-editor/lib/operations/factory-graph-operations";
 import {
   EXHAUSTION_WORKSTATION_ICON_METADATA,
   SUPPORTED_WORKSTATION_ICON_METADATA,
 } from "../../flowchart/lib/workstation-icon-metadata";
 import type { ReadFactoryImportFile } from "../../import/hooks/use-factory-png-drop";
-import { createFactoryImportConfirmInput } from "../../import/lib/factory-import-confirm-input.test-helpers";
 import type { FactoryImportConfirmInput } from "../../import/lib/factory-import-save-choice";
-import type { FactoryPngImportValue } from "../../import/lib/factory-png-import";
 import { getImportPreviewDialogMessages } from "../../import/messages/import-preview-dialog";
-import type { CurrentActivityImportController } from "../hooks/current-activity-import-controller";
-import { resetCurrentActivityGraphLayoutCacheForTests } from "../hooks/react-flow-current-activity-card-graph-layout";
 import {
   buildCurrentActivityGraphLayoutFromFactory,
   dashboardWorkstationFromFactory,
@@ -80,82 +63,21 @@ import {
   currentActivityGraphKey,
   currentActivityTopologyKey,
 } from "../lib/react-flow-current-activity-card-keys";
-import type { CurrentActivitySelection } from "../lib/react-flow-current-activity-card-types";
 import { getDashboardFlowAxisLegendMessages } from "../messages/dashboard-flow-axis-legend";
 import { getWorkflowActivityGraphImportMessages } from "../messages/graph-import";
 import { ReactFlowCurrentActivityCard } from "./react-flow-current-activity-card";
-
-vi.mock(
-  "../../current-factory-definition/hooks/useCurrentFactoryDefinition",
-  async () => {
-    const actual = await vi.importActual(
-      "../../current-factory-definition/hooks/useCurrentFactoryDefinition",
-    );
-
-    return {
-      ...actual,
-      useCurrentFactoryDocument: vi.fn(),
-    };
-  },
-);
-
-vi.mock(
-  "../../current-factory-definition/hooks/useFactoryDocumentSave",
-  () => ({
-    useFactoryDocumentSave: vi.fn(),
-  }),
-);
-
-vi.mock(
-  "../../factory-graph-editor/hooks/use-editable-factory-graph",
-  async () => {
-    const actual = await vi.importActual(
-      "../../factory-graph-editor/hooks/use-editable-factory-graph",
-    );
-
-    return {
-      ...actual,
-      useEditableFactoryGraph: vi.fn(),
-    };
-  },
-);
-
-vi.mock(
-  "../../factory-graph-editor/hooks/factory-graph-draft-hook",
-  async () => {
-    const actual = await vi.importActual(
-      "../../factory-graph-editor/hooks/factory-graph-draft-hook",
-    );
-
-    return {
-      ...actual,
-      useFactoryGraphDraftState: vi.fn(),
-    };
-  },
-);
-
-vi.mock(
-  "../../../components/ui/dialog",
-  () => import("../../../testing/mock-dashboard-dialog"),
-);
-
-const PADDING_CLASS_PATTERN = /(^|\s)p[trblxy]?-[^\s]+/;
-
-interface RenderCurrentActivityOptions {
-  activateFactory?: (
-    input: FactoryImportConfirmInput,
-  ) => Promise<ImportFactoryValue>;
-  currentFactoryDocument?: typeof baseFactoryDefinitionDocument | null;
-  currentFactoryDocumentStatus?: "error" | "pending" | "success";
-  importController?: CurrentActivityImportController;
-  locale?: string;
-  onFactoryActivated?: () => void;
-  onFactoryImportReady?: (value: FactoryPngImportValue, file: File) => void;
-  readFactoryImportFile?: ReadFactoryImportFile;
-  snapshot: DashboardSnapshot;
-  selection?: CurrentActivitySelection | null;
-  widgetInstanceID?: string;
-}
+import {
+  createFactoryImportValue,
+  createFileDropTransfer,
+  dashboardSnapshotWithActiveWorkItemCount,
+  defaultDraftState,
+  refreshFactoryFromTopology,
+  registerCurrentActivityCardTestLifecycle,
+  reinstallDashboardBrowserTestShims,
+  renderCurrentActivity,
+  renderWithQueryClient,
+  workerDenseSnapshot,
+} from "./react-flow-current-activity-card-component.harness.test";
 
 const LEGEND_ICON_EXPECTATIONS = [
   ["Queue", "queue"],
@@ -177,8 +99,6 @@ const LEGEND_ICON_EXPECTATIONS = [
 ] as const;
 const workflowGraphLocaleFallbackTimeoutMs = 180_000;
 
-const defaultDraftState = createMockGraphEditorDraftState();
-
 function dashboardSnapshotWithStateCounts(
   overrides: Record<string, number>,
 ): DashboardSnapshot {
@@ -186,78 +106,6 @@ function dashboardSnapshotWithStateCounts(
   snapshot.runtime.place_token_counts = {
     ...snapshot.runtime.place_token_counts,
     ...overrides,
-  };
-
-  return snapshot;
-}
-
-function refreshFactoryFromTopology(
-  snapshot: DashboardSnapshot,
-): DashboardSnapshot {
-  snapshot.factory = factoryFromDashboardTopology(snapshot.topology);
-  return snapshot;
-}
-
-function currentFactoryDocumentFromSnapshot(
-  snapshot: DashboardSnapshot,
-): typeof baseFactoryDefinitionDocument {
-  if (!snapshot.factory) {
-    return baseFactoryDefinitionDocument;
-  }
-
-  return {
-    ...snapshot.factory,
-    version: baseFactoryDefinitionDocument.version,
-  };
-}
-
-function dashboardSnapshotWithEditableFactory(): DashboardSnapshot {
-  const snapshot = structuredClone(semanticWorkflowDashboardSnapshot);
-  snapshot.factory = structuredClone(baseFactoryDefinitionDocument);
-  return snapshot;
-}
-
-function workerDenseSnapshot(): DashboardSnapshot {
-  const snapshot = structuredClone(semanticWorkflowDashboardSnapshot);
-  snapshot.runtime.active_executions_by_dispatch_id = {
-    "dispatch-draft-active": {
-      dispatch_id: "dispatch-draft-active",
-      started_at: "2026-05-19T01:10:00Z",
-      transition_id: "draft-transition",
-      workstation_name: "draft",
-      workstation_node_id: "draft",
-    },
-  };
-  snapshot.runtime.active_dispatch_ids = ["dispatch-draft-active"];
-  snapshot.runtime.active_workstation_node_ids = ["draft"];
-  snapshot.runtime.active_throttle_pauses = [
-    {
-      affected_worker_types: ["stalled"],
-      lane_id: "provider:codex",
-      model: "gpt-5",
-      paused_until: "2026-05-19T01:15:00Z",
-      provider: "OPENAI",
-      recover_at: "2026-05-19T01:16:00Z",
-    },
-  ];
-  snapshot.runtime.workstation_requests_by_dispatch_id = {
-    "dispatch-review": {
-      counts: {
-        dispatched_count: 1,
-        errored_count: 1,
-        responded_count: 1,
-      },
-      dispatch_id: "dispatch-review",
-      request: {},
-      response: {
-        failure_message: "Provider request failed.",
-        outcome: "FAILED",
-      },
-      transition_id: "review-transition",
-      workstation_name: "review",
-      workstation_node_id: "review",
-      work_items: [],
-    },
   };
 
   return snapshot;
@@ -432,144 +280,6 @@ async function expectRenderableCurrentActivityGraphEdges(
     ),
   ).toBe(true);
 }
-
-function renderCurrentActivity({
-  activateFactory,
-  currentFactoryDocument,
-  currentFactoryDocumentStatus = "success",
-  importController,
-  locale,
-  onFactoryActivated,
-  onFactoryImportReady,
-  readFactoryImportFile,
-  snapshot,
-  selection = null,
-  widgetInstanceID,
-}: RenderCurrentActivityOptions) {
-  const factoryDocument =
-    currentFactoryDocument !== undefined
-      ? currentFactoryDocument
-      : currentFactoryDocumentFromSnapshot(snapshot);
-
-  vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-    data: factoryDocument ?? undefined,
-    error: null,
-    status: currentFactoryDocumentStatus,
-  } as never);
-
-  const onSelectWorkID =
-    vi.fn<
-      (workID: string, hint?: { dispatchID?: string; nodeID?: string }) => void
-    >();
-  const onSelectStateNode = vi.fn<(placeId: string) => void>();
-  const onSelectDoc = vi.fn<(targetPath: string) => void>();
-  const onSelectResource = vi.fn<(resourceName: string) => void>();
-  const onSelectWorker = vi.fn<(workerName: string) => void>();
-  const onSelectWorkType = vi.fn<(workTypeName: string) => void>();
-  const onSelectWorkstation = vi.fn<(nodeId: string) => void>();
-
-  renderWithQueryClient(
-    <ReactFlowCurrentActivityCard
-      activateFactory={activateFactory}
-      importController={importController}
-      locale={locale}
-      now={Date.parse("2026-04-08T12:00:04Z")}
-      onFactoryActivated={onFactoryActivated}
-      onFactoryImportReady={onFactoryImportReady}
-      onSelectWorkID={onSelectWorkID}
-      onSelectDoc={onSelectDoc}
-      onSelectResource={onSelectResource}
-      onSelectStateNode={onSelectStateNode}
-      onSelectWorker={onSelectWorker}
-      onSelectWorkType={onSelectWorkType}
-      onSelectWorkstation={onSelectWorkstation}
-      readFactoryImportFile={readFactoryImportFile}
-      selection={selection}
-      snapshot={snapshot}
-      widgetInstanceID={widgetInstanceID}
-    />,
-  );
-
-  return {
-    onSelectDoc,
-    onSelectResource,
-    onSelectStateNode,
-    onSelectWorkID,
-    onSelectWorker,
-    onSelectWorkType,
-    onSelectWorkstation,
-  };
-}
-
-function renderWithQueryClient(view: ReactElement) {
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      mutations: {
-        retry: false,
-      },
-      queries: {
-        gcTime: Infinity,
-        retry: false,
-      },
-    },
-  });
-
-  return render(
-    <QueryClientProvider client={queryClient}>
-      <DashboardSessionTestProvider>{view}</DashboardSessionTestProvider>
-    </QueryClientProvider>,
-  );
-}
-
-function createFactoryImportValue(): FactoryPngImportValue {
-  return {
-    factory: {
-      name: "Dropped Factory",
-      workTypes: [],
-      workers: [],
-      workstations: [],
-    },
-    previewImageSrc: "blob:factory-preview",
-    revokePreviewImageSrc: vi.fn(),
-    schemaVersion: "portos.agent-factory.png.v1",
-  };
-}
-
-function createImportController(
-  overrides: Partial<CurrentActivityImportController> = {},
-): CurrentActivityImportController {
-  return {
-    activateImport: vi.fn().mockResolvedValue(undefined),
-    activationState: { status: "idle" },
-    clearActivationError: vi.fn(),
-    clearError: vi.fn(),
-    closeImportPreview: vi.fn(),
-    dropState: { status: "idle" },
-    importPreviewState: { status: "idle" },
-    onDragEnter: vi.fn(),
-    onDragLeave: vi.fn(),
-    onDragOver: vi.fn(),
-    onDrop: vi.fn(),
-    ...overrides,
-  };
-}
-
-function createFileDropTransfer(files: File[]): {
-  dataTransfer: {
-    dropEffect: string;
-    files: File[];
-    types: string[];
-  };
-} {
-  return {
-    dataTransfer: {
-      dropEffect: "none",
-      files,
-      types: ["Files"],
-    },
-  };
-}
-
 async function expandGraphLegend(locale = "en"): Promise<HTMLElement> {
   const messages = getDashboardFlowAxisLegendMessages(locale);
   const actionTargetLabel =
@@ -650,72 +360,6 @@ function dashboardSnapshotWithLongStateLabels(): DashboardSnapshot {
   }
 
   return refreshFactoryFromTopology(snapshot);
-}
-
-function dashboardSnapshotWithActiveWorkItemCount(
-  count: number,
-): DashboardSnapshot {
-  const snapshot = structuredClone(semanticWorkflowDashboardSnapshot);
-  const reviewActivity =
-    snapshot.runtime.workstation_activity_by_node_id?.review;
-  const activeExecution =
-    snapshot.runtime.active_executions_by_dispatch_id?.[
-      "dispatch-review-active"
-    ];
-  const workItems = Array.from(
-    { length: count },
-    (_, index): DashboardWorkItemRef => {
-      const itemNumber = index + 1;
-
-      return {
-        display_name: `Active Story ${itemNumber}`,
-        trace_id: `trace-active-story-${itemNumber}`,
-        work_id: `work-active-story-${itemNumber}`,
-        work_type_id: "story",
-      };
-    },
-  );
-
-  if (count === 0) {
-    snapshot.runtime.active_dispatch_ids = (
-      snapshot.runtime.active_dispatch_ids ?? []
-    ).filter((dispatchID) => dispatchID !== "dispatch-review-active");
-    snapshot.runtime.active_workstation_node_ids = (
-      snapshot.runtime.active_workstation_node_ids ?? []
-    ).filter((nodeID) => nodeID !== "review");
-    if (snapshot.runtime.active_executions_by_dispatch_id) {
-      delete snapshot.runtime.active_executions_by_dispatch_id[
-        "dispatch-review-active"
-      ];
-    }
-    if (reviewActivity) {
-      reviewActivity.active_dispatch_ids = [];
-      reviewActivity.active_work_items = [];
-      reviewActivity.trace_ids = [];
-    }
-  } else if (activeExecution) {
-    activeExecution.work_items = workItems;
-    activeExecution.trace_ids = workItems.map(
-      (workItem) => workItem.trace_id ?? workItem.work_id,
-    );
-    snapshot.runtime.active_dispatch_ids = ["dispatch-review-active"];
-    snapshot.runtime.active_workstation_node_ids = ["review"];
-    if (reviewActivity) {
-      reviewActivity.active_dispatch_ids = ["dispatch-review-active"];
-      reviewActivity.active_work_items = workItems;
-      reviewActivity.trace_ids = workItems.map(
-        (workItem) => workItem.trace_id ?? workItem.work_id,
-      );
-    }
-  }
-
-  snapshot.runtime.in_flight_dispatch_count = count;
-
-  const refreshedSnapshot = refreshFactoryFromTopology(snapshot);
-  if (refreshedSnapshot.factory) {
-    refreshedSnapshot.factory.version = baseFactoryDefinitionDocument.version;
-  }
-  return refreshedSnapshot;
 }
 
 function dashboardSnapshotWithActiveImplementWorkstation(): DashboardSnapshot {
@@ -956,1660 +600,6 @@ function dashboardSnapshotWithExhaustionRuleNode(): DashboardSnapshot {
 
   return snapshot;
 }
-
-let restoreBrowserTestShims: (() => void) | null = null;
-
-function registerCurrentActivityCardTestLifecycle(): void {
-  beforeEach(() => {
-    window.localStorage.clear();
-    resetCurrentActivityGraphLayoutCacheForTests();
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
-    vi.mocked(useFactoryDocumentSave).mockReturnValue({
-      error: null,
-      isPending: false,
-      reset: vi.fn(),
-      save: vi.fn(),
-      saveAsync: vi.fn(),
-    } as never);
-    wireMockEditableFactoryGraph(
-      {
-        useEditableFactoryGraph: vi.mocked(useEditableFactoryGraph),
-        useFactoryGraphDraftState: vi.mocked(useFactoryGraphDraftState),
-      },
-      defaultDraftState,
-    );
-  });
-
-  afterEach(() => {
-    cleanup();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = null;
-    vi.clearAllMocks();
-  });
-}
-
-function registerCurrentActivityCardEditorChromeTests(): void {
-  it("keeps editor controls unavailable until the graph editor mode is enabled", async () => {
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    expect(screen.getByRole("button", { name: "Edit mode" })).toBeTruthy();
-    const toolbar = screen.getByRole("region", {
-      name: "Factory graph editor tools",
-    });
-    expect(
-      within(toolbar).getByRole("button", {
-        name: "Show or hide",
-      }),
-    ).toBeTruthy();
-    expect(within(toolbar).queryByRole("button", { name: "Add" })).toBeNull();
-    expect(screen.queryByText("Observe")).toBeNull();
-  });
-
-  it("shows the add and delete toolbar in editor mode", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    const toolbar = await screen.findByRole("region", {
-      name: "Factory graph editor tools",
-    });
-    expect(within(toolbar).getByRole("button", { name: "Add" })).toBeTruthy();
-    expect(
-      within(toolbar).getByRole("button", {
-        name: "Delete, no graph items selected",
-      }),
-    ).toBeTruthy();
-    expect(screen.queryByText("Editor mode active")).toBeNull();
-  });
-
-  it("keeps classifier workstations out of the editable graph flow", async () => {
-    const snapshot = dashboardSnapshotWithActiveWorkItemCount(0);
-    snapshot.topology.workstation_nodes_by_id.review.workstation_kind =
-      "CLASSIFIER_WORKSTATION";
-    snapshot.factory = structuredClone(baseFactoryDefinition);
-    const reviewWorkstation = snapshot.factory?.workstations?.find(
-      (workstation) => workstation.name === "review",
-    );
-    if (reviewWorkstation) {
-      reviewWorkstation.type = "CLASSIFIER_WORKSTATION";
-      reviewWorkstation.classificationRoutes = [
-        {
-          label: "approved",
-          output: {
-            state: "complete",
-            workType: "story",
-          },
-        },
-      ];
-    }
-
-    renderCurrentActivity({
-      currentFactoryDocument: null,
-      currentFactoryDocumentStatus: "pending",
-      snapshot,
-    });
-
-    const unavailableEditorButton = screen.getByRole("button", {
-      name: 'Factory graph editing does not yet support classifier workstation routes. "review" stays read-only in this view until labeled route editing is available.',
-    });
-    expect(unavailableEditorButton.getAttribute("disabled")).not.toBeNull();
-    expect(
-      screen.queryByText(
-        'Editor unavailable: Factory graph editing does not yet support classifier workstation routes. "review" stays read-only in this view until labeled route editing is available.',
-      ),
-    ).toBeNull();
-
-    const toolbar = screen.getByRole("region", {
-      name: "Factory graph editor tools",
-    });
-    expect(
-      within(toolbar).getByRole("button", {
-        name: "Show or hide",
-      }),
-    ).toBeTruthy();
-    expect(within(toolbar).queryByRole("button", { name: "Add" })).toBeNull();
-    expect(screen.queryByText("Editor mode active")).toBeNull();
-  });
-
-  it("lists the supported add-entity options and validates duplicate worker names before mutating the draft", async () => {
-    const replaceDraft = vi.fn();
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      replaceDraft,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(await screen.findByRole("button", { name: "Add" }));
-
-    expect(screen.getByRole("button", { name: "Workstation" })).toBeTruthy();
-    expect(screen.getByRole("button", { name: "Worker" })).toBeTruthy();
-    expect(screen.getByRole("button", { name: "Work type" })).toBeTruthy();
-    expect(screen.getByRole("button", { name: "Work state" })).toBeTruthy();
-    expect(screen.getByRole("button", { name: "Resource" })).toBeTruthy();
-
-    fireEvent.click(screen.getByRole("button", { name: "Worker" }));
-    fireEvent.change(screen.getByRole("textbox", { name: "Identifier" }), {
-      target: { value: "writer" },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "Add entity" }));
-
-    expect(
-      screen.getByText('A worker named "writer" already exists in the draft.'),
-    ).toBeTruthy();
-    expect(replaceDraft).not.toHaveBeenCalled();
-  });
-
-  it("submits valid add-entity forms into the pending graph draft", async () => {
-    const replaceDraft = vi.fn();
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      replaceDraft,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(await screen.findByRole("button", { name: "Add" }));
-    fireEvent.click(screen.getByRole("button", { name: "Work type" }));
-    fireEvent.change(screen.getByRole("textbox", { name: "Identifier" }), {
-      target: { value: "essay" },
-    });
-    fireEvent.change(screen.getByRole("textbox", { name: "First state" }), {
-      target: { value: "queued" },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "Add entity" }));
-
-    expect(replaceDraft).toHaveBeenCalledTimes(1);
-    const nextDraft = replaceDraft.mock.calls[0]?.[0];
-    expect(nextDraft.additions.workTypes).toEqual([
-      {
-        name: "essay",
-        states: [
-          {
-            name: "queued",
-            type: "INITIAL",
-          },
-        ],
-      },
-    ]);
-  });
-
-  it("distinguishes work-state creation from work-type creation and blocks missing work-type association", async () => {
-    const replaceDraft = vi.fn();
-    const user = userEvent.setup();
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      replaceDraft,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(await screen.findByRole("button", { name: "Add" }));
-    fireEvent.click(screen.getByRole("button", { name: "Work state" }));
-
-    expect(screen.getByText("Add work state")).toBeTruthy();
-    expect(
-      screen.getByText("Append a new ordered state to an existing work type."),
-    ).toBeTruthy();
-    expect(
-      screen.getByRole("combobox", { name: "Work type" }),
-    ).toHaveTextContent("story");
-    expect(
-      screen.getByRole("combobox", { name: "State type" }),
-    ).toHaveTextContent("PROCESSING");
-
-    await selectLabeledComboboxOption(user, "Work type", "Select a work type");
-    fireEvent.click(screen.getByRole("button", { name: "Add entity" }));
-
-    expect(
-      screen.getByText("Choose a work type before adding a work state."),
-    ).toBeTruthy();
-    expect(replaceDraft).not.toHaveBeenCalled();
-  });
-
-  it("submits valid work-state add-entity forms into the pending graph draft", async () => {
-    const replaceDraft = vi.fn();
-    const user = userEvent.setup();
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      replaceDraft,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithEditableFactory(),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(await screen.findByRole("button", { name: "Add" }));
-    fireEvent.click(screen.getByRole("button", { name: "Work state" }));
-    fireEvent.change(screen.getByRole("textbox", { name: "Identifier" }), {
-      target: { value: "approved" },
-    });
-    await selectLabeledComboboxOption(user, "State type", "TERMINAL");
-    fireEvent.click(screen.getByRole("button", { name: "Add entity" }));
-
-    expect(replaceDraft).toHaveBeenCalledTimes(1);
-    const nextDraft = replaceDraft.mock.calls[0]?.[0];
-
-    expect(nextDraft.additions.workStates).toEqual([
-      {
-        state: {
-          name: "approved",
-          type: "TERMINAL",
-        },
-        workTypeName: "story",
-      },
-    ]);
-  }, 30_000);
-
-  it("keeps editor mode on the shared observer graph surface", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      graph: {
-        edges: [],
-        nodes: [
-          {
-            id: "worker:writer",
-            key: { kind: "worker", name: "writer" },
-            kind: "worker",
-            label: "writer",
-          },
-          {
-            id: "workstation:review",
-            key: { kind: "workstation", name: "review" },
-            kind: "workstation",
-            label: "review",
-          },
-        ],
-      },
-      hasChanges: true,
-      draft: {
-        ...defaultDraftState.draft,
-        additions: {
-          ...defaultDraftState.draft.additions,
-          workstations: [
-            {
-              inputs: [],
-              name: "review",
-              outputs: [],
-              type: "MODEL_WORKSTATION",
-              worker: "writer",
-            },
-          ],
-        },
-      },
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    await waitFor(() => {
-      expect(
-        document.querySelector(
-          '[data-current-activity-node-type="workstation"]',
-        ),
-      ).toBeTruthy();
-    });
-    expect(screen.queryByText("Pending")).toBeNull();
-  });
-
-  it("renders worker and resource nodes from the canonical snapshot factory in observer mode", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: workerDenseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      latestDocument: workerDenseFactoryDefinitionDocument,
-      pendingFactoryDefinition: workerDenseFactoryDefinitionDocument,
-    } as never);
-    const snapshot = workerDenseSnapshot();
-    snapshot.factory = workerDenseFactoryDefinitionDocument;
-
-    renderCurrentActivity({
-      snapshot,
-    });
-
-    await waitFor(() => {
-      expect(
-        document.querySelector(
-          '[data-current-activity-node-type="workstation"]',
-        ),
-      ).toBeTruthy();
-    });
-
-    await waitFor(() => {
-      expect(document.querySelector('[data-id="worker:writer"]')).toBeTruthy();
-      expect(
-        document.querySelector('[data-id="worker:reviewer"]'),
-      ).toBeTruthy();
-      expect(document.querySelector('[data-id="worker:stalled"]')).toBeTruthy();
-      expect(document.querySelector('[data-id="resource:gpu"]')).toBeTruthy();
-    });
-  });
-
-  it("renders event snapshot workstations in observe mode while the factory document is pending", async () => {
-    const snapshot = buildDivergentPlaneDashboardSnapshot();
-    if (snapshot.factory) {
-      snapshot.factory.workstations = snapshot.factory.workstations?.filter(
-        (workstation) => workstation.name !== "Document Only",
-      );
-    }
-    wireMockEditableFactoryGraph(
-      {
-        useEditableFactoryGraph: vi.mocked(useEditableFactoryGraph),
-        useFactoryGraphDraftState: vi.mocked(useFactoryGraphDraftState),
-      },
-      createMockGraphEditorDraftState({
-        baseDocument: null,
-        latestDocument: null,
-        pendingFactoryDefinition: null,
-      }),
-    );
-
-    renderCurrentActivity({
-      currentFactoryDocument: null,
-      currentFactoryDocumentStatus: "pending",
-      snapshot,
-    });
-
-    expect(
-      await screen.findByRole("region", { name: "Work graph viewport" }),
-    ).toBeTruthy();
-    expect(
-      await screen.findByRole("button", {
-        name: "Select Snapshot Only workstation",
-      }),
-    ).toBeTruthy();
-    expect(
-      screen.queryByRole("button", {
-        name: "Select Document Only workstation",
-      }),
-    ).toBeNull();
-  });
-
-  it("renders nested bundled docs as observe-mode graph nodes from the saved factory document", async () => {
-    const snapshot = structuredClone(semanticWorkflowDashboardSnapshot);
-    refreshFactoryFromTopology(snapshot);
-    const nestedDocPath = "factory/docs/standards/review.md";
-    const savedDocument = {
-      ...currentFactoryDocumentFromSnapshot(snapshot),
-      supportingFiles: {
-        bundledFiles: [
-          {
-            content: { encoding: "utf-8", inline: "# Overview" },
-            targetPath: "factory/docs/overview.md",
-            type: "DOC",
-          },
-          {
-            content: { encoding: "utf-8", inline: "# Review standards" },
-            targetPath: nestedDocPath,
-            type: "DOC",
-          },
-        ],
-      },
-    };
-    snapshot.factory = savedDocument;
-
-    renderCurrentActivity({
-      snapshot,
-    });
-
-    await waitFor(() => {
-      expect(
-        screen.getByRole("button", { name: "Select review.md doc" }),
-      ).toBeTruthy();
-    });
-    expect(screen.getByText(nestedDocPath)).toBeTruthy();
-  });
-
-  it("renders bundled docs as observe-mode graph nodes from the saved factory document", async () => {
-    const snapshot = structuredClone(semanticWorkflowDashboardSnapshot);
-    refreshFactoryFromTopology(snapshot);
-    const savedDocument = {
-      ...currentFactoryDocumentFromSnapshot(snapshot),
-      supportingFiles: {
-        bundledFiles: [
-          {
-            content: { encoding: "utf-8", inline: "# Overview" },
-            targetPath: "factory/docs/overview.md",
-            type: "DOC",
-          },
-          {
-            content: { encoding: "utf-8", inline: "# Planning" },
-            targetPath: "factory/docs/planning.md",
-            type: "DOC",
-          },
-        ],
-      },
-    };
-    snapshot.factory = savedDocument;
-
-    renderCurrentActivity({
-      snapshot,
-    });
-
-    await waitFor(() => {
-      expect(
-        screen.getByRole("button", { name: "Select overview.md doc" }),
-      ).toBeTruthy();
-    });
-    expect(
-      screen.getByRole("button", { name: "Select planning.md doc" }),
-    ).toBeTruthy();
-  });
-
-  it("renders event snapshot workstations in observe mode when the document plane diverges from the snapshot", async () => {
-    const snapshot = buildDivergentPlaneDashboardSnapshot();
-    if (snapshot.factory) {
-      snapshot.factory.workstations = snapshot.factory.workstations?.filter(
-        (workstation) => workstation.name !== "Document Only",
-      );
-    }
-    wireMockEditableFactoryGraph(
-      {
-        useEditableFactoryGraph: vi.mocked(useEditableFactoryGraph),
-        useFactoryGraphDraftState: vi.mocked(useFactoryGraphDraftState),
-      },
-      createMockGraphEditorDraftState({
-        baseDocument: divergentDocumentPlaneFactoryDocument,
-        latestDocument: divergentDocumentPlaneFactoryDocument,
-        pendingFactoryDefinition: divergentDocumentPlaneFactoryDocument,
-      }),
-    );
-
-    renderCurrentActivity({
-      currentFactoryDocument: divergentDocumentPlaneFactoryDocument,
-      snapshot,
-    });
-
-    expect(
-      await screen.findByRole("button", {
-        name: "Select Snapshot Only workstation",
-      }),
-    ).toBeTruthy();
-    expect(
-      screen.queryByRole("button", {
-        name: "Select Document Only workstation",
-      }),
-    ).toBeNull();
-  });
-
-  it("renders document-only workstations in edit mode when the snapshot plane diverges", async () => {
-    const snapshot = buildDivergentPlaneDashboardSnapshot();
-    wireMockEditableFactoryGraph(
-      {
-        useEditableFactoryGraph: vi.mocked(useEditableFactoryGraph),
-        useFactoryGraphDraftState: vi.mocked(useFactoryGraphDraftState),
-      },
-      createMockGraphEditorDraftState({
-        baseDocument: divergentDocumentPlaneFactoryDocument,
-        latestDocument: divergentDocumentPlaneFactoryDocument,
-        pendingFactoryDefinition: divergentDocumentPlaneFactoryDocument,
-      }),
-    );
-
-    renderCurrentActivity({
-      currentFactoryDocument: divergentDocumentPlaneFactoryDocument,
-      snapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    await waitFor(() => {
-      expect(
-        screen.getByRole("button", {
-          name: "Select Document Only workstation",
-        }),
-      ).toBeTruthy();
-    });
-    expect(
-      screen.queryByRole("button", {
-        name: "Select Snapshot Only workstation",
-      }),
-    ).toBeNull();
-  });
-
-  it("does not render snapshot-only workstations in editor mode while the factory document is loading", async () => {
-    const snapshot = buildDivergentPlaneDashboardSnapshot();
-    wireMockEditableFactoryGraph(
-      {
-        useEditableFactoryGraph: vi.mocked(useEditableFactoryGraph),
-        useFactoryGraphDraftState: vi.mocked(useFactoryGraphDraftState),
-      },
-      createMockGraphEditorDraftState({
-        baseDocument: null,
-        latestDocument: null,
-        pendingFactoryDefinition: null,
-      }),
-    );
-
-    renderCurrentActivity({
-      currentFactoryDocument: null,
-      currentFactoryDocumentStatus: "pending",
-      snapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    await waitFor(() => {
-      expect(
-        screen.getByRole("region", { name: "Factory graph editor tools" }),
-      ).toBeTruthy();
-    });
-    expect(
-      screen.queryByRole("button", {
-        name: "Select Snapshot Only workstation",
-      }),
-    ).toBeNull();
-    expect(
-      screen.queryByRole("button", {
-        name: "Select Document Only workstation",
-      }),
-    ).toBeNull();
-  });
-
-  it("does not render the editor-only visibility preset controls in embedded editor mode", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: workerDenseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      latestDocument: workerDenseFactoryDefinitionDocument,
-      pendingFactoryDefinition: workerDenseFactoryDefinitionDocument,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: workerDenseSnapshot(),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    await waitFor(() => {
-      expect(
-        document.querySelector(
-          '[data-current-activity-node-type="workstation"]',
-        ),
-      ).toBeTruthy();
-    });
-
-    expect(screen.queryByRole("button", { name: "Infrastructure" })).toBeNull();
-    expect(screen.queryByRole("button", { name: "Workflow" })).toBeNull();
-    expect(screen.queryByRole("button", { name: "All" })).toBeNull();
-  });
-
-  it("renders supported workstation and work-state editor handles on the shared observer graph", async () => {
-    const snapshot = semanticWorkflowDashboardSnapshot;
-    const document = currentFactoryDocumentFromSnapshot(snapshot);
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: document,
-      error: null,
-      status: "success",
-    } as never);
-    wireMockEditableFactoryGraph(
-      {
-        useEditableFactoryGraph: vi.mocked(useEditableFactoryGraph),
-        useFactoryGraphDraftState: vi.mocked(useFactoryGraphDraftState),
-      },
-      createMockGraphEditorDraftState({
-        baseDocument: document,
-        latestDocument: document,
-        pendingFactoryDefinition: document,
-      }),
-    );
-
-    renderCurrentActivity({
-      snapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    expect(
-      await screen.findAllByLabelText(
-        "Route successful output from this workstation.",
-      ),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText(
-        "Accept an input work state for this workstation.",
-      ),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText(
-        "Route this work state into a workstation input.",
-      ),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText(
-        "Receive workstation output into this work state.",
-      ),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText(
-        "Route successful output from this workstation.",
-      ),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText("Assign this worker to a workstation."),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText(
-        "Accept a worker assignment for this workstation.",
-      ),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText(
-        "Accept a resource requirement for this workstation.",
-      ),
-    ).not.toHaveLength(0);
-    expect(
-      screen.getAllByLabelText("Provide this resource to a workstation."),
-    ).not.toHaveLength(0);
-  });
-}
-
-function registerCurrentActivityCardEditorLeaveAndSaveTests(): void {
-  it("removes a workstation without opening a confirmation", () => {
-    const result = removeFactoryGraphNode({
-      baseFactoryDefinition: baseFactoryDefinitionDocument,
-      draft: defaultDraftState.draft,
-      nodeId: "workstation:review",
-    });
-
-    expect(result.ok).toBe(true);
-    if (!result.ok) {
-      return;
-    }
-    expect(result.value.removals.workstations).toEqual(["review"]);
-  });
-
-  it("keeps worker nodes visible but not workstation-style deletion targets", async () => {
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithEditableFactory(),
-    });
-
-    expect(
-      await screen.findByRole("button", {
-        name: "Select writer worker",
-      }),
-    ).toBeTruthy();
-    expect(await screen.findByLabelText("worker:writer")).toBeTruthy();
-    expect(
-      await screen.findByRole("button", {
-        name: /Select .* workstation/,
-      }),
-    ).toBeTruthy();
-
-    const removeWorker = removeFactoryGraphNode({
-      baseFactoryDefinition: baseFactoryDefinitionDocument,
-      draft: defaultDraftState.draft,
-      nodeId: "worker:writer",
-    });
-    expect(removeWorker).toMatchObject({
-      ok: false,
-      reason: "BLOCKED_REMOVAL",
-    });
-  });
-
-  it("keeps removed server-backed workstations visible with a pending-removal badge", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      draft: {
-        ...defaultDraftState.draft,
-        removals: {
-          ...defaultDraftState.draft.removals,
-          workstations: ["review"],
-        },
-      },
-      graph: {
-        edges: [],
-        nodes: [],
-      },
-      hasChanges: true,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    expect(
-      await screen.findByRole("button", {
-        name: "Select Review workstation",
-      }),
-    ).toBeTruthy();
-    const toggle = screen.getByRole("button", {
-      name: "Leave editor",
-    });
-    expect(toggle.className).toContain("border-af-warning-border");
-  });
-
-  it("shows a loading editor state while the editable definition is still fetching", async () => {
-    renderCurrentActivity({
-      currentFactoryDocument: null,
-      currentFactoryDocumentStatus: "pending",
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    const toolbar = await screen.findByRole("region", {
-      name: "Factory graph editor tools",
-    });
-    expect(screen.queryByText("Loading editor definition")).toBeNull();
-    expect(
-      within(toolbar)
-        .getByRole("button", { name: "Add" })
-        .getAttribute("disabled"),
-    ).not.toBeNull();
-    expect(
-      within(toolbar)
-        .getByRole("button", { name: "Delete, no graph items selected" })
-        .getAttribute("disabled"),
-    ).not.toBeNull();
-  });
-
-  it("requires save, discard, or keep editing before leaving editor mode with unsaved changes", async () => {
-    const resetDraft = vi.fn();
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      hasChanges: true,
-      resetDraft,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(screen.getByRole("button", { name: "Leave editor" }));
-
-    const dialog = await screen.findByRole("dialog", {
-      name: "Leave graph editor with unsaved changes?",
-    });
-    expect(
-      within(dialog).getByRole("button", { name: "Save changes" }),
-    ).toBeTruthy();
-    expect(
-      within(dialog).getByRole("button", { name: "Discard changes" }),
-    ).toBeTruthy();
-    expect(
-      within(dialog).getByRole("button", { name: "Keep editing" }),
-    ).toBeTruthy();
-
-    fireEvent.click(
-      within(dialog).getByRole("button", { name: "Discard changes" }),
-    );
-
-    expect(resetDraft).toHaveBeenCalledTimes(1);
-  });
-
-  it("keeps editor mode when choosing Keep editing on the leave dialog", async () => {
-    const resetDraft = vi.fn();
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      hasChanges: true,
-      resetDraft,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(screen.getByRole("button", { name: "Leave editor" }));
-
-    const dialog = await screen.findByRole("dialog", {
-      name: "Leave graph editor with unsaved changes?",
-    });
-    fireEvent.click(
-      within(dialog).getByRole("button", { name: "Keep editing" }),
-    );
-
-    await waitFor(() => {
-      expect(screen.queryByRole("dialog")).toBeNull();
-    });
-    expect(resetDraft).not.toHaveBeenCalled();
-    expect(
-      screen
-        .getByRole("button", { name: "Leave editor" })
-        .getAttribute("aria-pressed"),
-    ).toBe("true");
-    expect(screen.queryByText("Unsaved changes")).toBeNull();
-  });
-
-  it("shows explicit save and discard actions for pending graph changes", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      draft: {
-        ...defaultDraftState.draft,
-        additions: {
-          ...defaultDraftState.draft.additions,
-          workers: [
-            {
-              model: "gpt-5-mini",
-              name: "reviewer",
-              type: "MODEL_WORKER",
-            },
-          ],
-        },
-      },
-      hasChanges: true,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    const actions = await screen.findByRole("region", {
-      name: "Factory graph editor tools",
-    });
-    expect(
-      within(actions).getByRole("button", { name: "Discard changes" }),
-    ).toBeTruthy();
-    expect(
-      within(actions).getByRole("button", { name: "Save changes" }),
-    ).toBeTruthy();
-    expect(
-      screen.queryByRole("region", { name: "Pending graph changes" }),
-    ).toBeNull();
-  });
-
-  it("shows consolidated unsaved chrome with a single header status and warning toggle", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      draft: {
-        ...defaultDraftState.draft,
-        additions: {
-          ...defaultDraftState.draft.additions,
-          workers: [
-            {
-              model: "gpt-5-mini",
-              name: "reviewer",
-              type: "MODEL_WORKER",
-            },
-          ],
-        },
-      },
-      hasChanges: true,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    const toolbar = await screen.findByRole("region", {
-      name: "Factory graph editor tools",
-    });
-    const statusSection = document.querySelector(
-      '[data-dashboard-action-row-section="statuses"]',
-    );
-    expect(statusSection).toBeNull();
-    expect(within(toolbar).queryAllByRole("status")).toHaveLength(0);
-
-    const toggle = screen.getByRole("button", {
-      name: "Leave editor",
-    });
-    expect(toggle.className).toContain("border-af-warning-border");
-    expect(toggle.className).toContain("bg-warning-container");
-    expect(toggle.className).toContain("text-on-warning-container");
-  });
-
-  it("confirms pending save changes before saving the graph draft", async () => {
-    const saveAsync = vi.fn().mockResolvedValue(baseFactoryDefinitionDocument);
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryDocumentSave).mockReturnValue({
-      error: null,
-      isPending: false,
-      reset: vi.fn(),
-      save: vi.fn(),
-      saveAsync,
-    } as never);
-    const draftWithReviewer = {
-      ...defaultDraftState,
-      draft: {
-        ...defaultDraftState.draft,
-        additions: {
-          ...defaultDraftState.draft.additions,
-          workers: [
-            {
-              model: "gpt-5-mini",
-              name: "reviewer",
-              type: "MODEL_WORKER",
-            },
-          ],
-        },
-      },
-      hasChanges: true,
-    };
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue(
-      draftWithReviewer as never,
-    );
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(
-      within(
-        await screen.findByRole("region", {
-          name: "Factory graph editor tools",
-        }),
-      ).getByRole("button", { name: "Save changes" }),
-    );
-
-    const dialog = await screen.findByRole("dialog", {
-      name: "Save factory graph changes?",
-    });
-    expect(
-      within(dialog).getByText("This save will apply 1 created entity."),
-    ).toBeTruthy();
-
-    fireEvent.click(
-      within(dialog).getByRole("button", { name: "Save topology" }),
-    );
-
-    await waitFor(() => {
-      expect(saveAsync).toHaveBeenCalledWith(
-        buildMockGraphSavePayload(draftWithReviewer),
-      );
-    });
-  });
-
-  it("warns when a newer editable-definition version arrives during a dirty draft", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      baseDocument: baseFactoryDefinitionDocument,
-      draft: {
-        ...defaultDraftState.draft,
-        additions: {
-          ...defaultDraftState.draft.additions,
-          workers: [
-            {
-              model: "gpt-5-mini",
-              name: "reviewer",
-              type: "MODEL_WORKER",
-            },
-          ],
-        },
-      },
-      hasChanges: true,
-      latestDocument: {
-        ...baseFactoryDefinitionDocument,
-        version: {
-          logical: "9",
-          physical: "2026-05-19T01:45:00Z",
-        },
-      },
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    expect(
-      await screen.findByText("A newer factory definition is available"),
-    ).toBeTruthy();
-    expect(
-      screen.getByText(
-        "Refresh or discard the current draft before saving so you do not overwrite a newer topology version.",
-      ),
-    ).toBeTruthy();
-    expect(
-      screen
-        .getByRole("button", { name: "Save changes" })
-        .getAttribute("disabled"),
-    ).not.toBeNull();
-  });
-
-  it("blocks topology save while active work is still running", async () => {
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue({
-      ...defaultDraftState,
-      draft: {
-        ...defaultDraftState.draft,
-        additions: {
-          ...defaultDraftState.draft.additions,
-          workers: [
-            {
-              model: "gpt-5-mini",
-              name: "reviewer",
-              type: "MODEL_WORKER",
-            },
-          ],
-        },
-      },
-      hasChanges: true,
-    } as never);
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(1),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-
-    expect(await screen.findByText("Topology edits are blocked")).toBeTruthy();
-    expect(
-      screen.getByText(
-        "Save is unavailable while active work is still running in the current factory.",
-      ),
-    ).toBeTruthy();
-    expect(
-      screen
-        .getByRole("button", { name: "Save changes" })
-        .getAttribute("disabled"),
-    ).not.toBeNull();
-  });
-
-  it("saves the pending editable definition before leaving editor mode", async () => {
-    const saveAsync = vi.fn().mockResolvedValue(baseFactoryDefinitionDocument);
-    const replaceDraft = vi.fn();
-    vi.mocked(useCurrentFactoryDocument).mockReturnValue({
-      data: baseFactoryDefinitionDocument,
-      error: null,
-      status: "success",
-    } as never);
-    vi.mocked(useFactoryDocumentSave).mockReturnValue({
-      error: null,
-      isPending: false,
-      reset: vi.fn(),
-      save: vi.fn(),
-      saveAsync,
-    } as never);
-    const dirtyDraftState = {
-      ...defaultDraftState,
-      hasChanges: true,
-      replaceDraft,
-    };
-    vi.mocked(useFactoryGraphDraftState).mockReturnValue(
-      dirtyDraftState as never,
-    );
-
-    renderCurrentActivity({
-      snapshot: dashboardSnapshotWithActiveWorkItemCount(0),
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Edit mode" }));
-    fireEvent.click(screen.getByRole("button", { name: "Leave editor" }));
-    fireEvent.click(
-      within(
-        await screen.findByRole("dialog", {
-          name: "Leave graph editor with unsaved changes?",
-        }),
-      ).getByRole("button", { name: "Save changes" }),
-    );
-
-    await waitFor(() => {
-      expect(saveAsync).toHaveBeenCalledWith(
-        buildMockGraphSavePayload(dirtyDraftState),
-      );
-    });
-    expect(replaceDraft).toHaveBeenCalledTimes(1);
-  });
-}
-
-describe("ReactFlowCurrentActivityCard editor chrome", () => {
-  registerCurrentActivityCardTestLifecycle();
-  registerCurrentActivityCardEditorChromeTests();
-  registerCurrentActivityCardEditorLeaveAndSaveTests();
-});
-
-describe("ReactFlowCurrentActivityCard import flows", () => {
-  registerCurrentActivityCardTestLifecycle();
-
-  it("scopes file drag-over and drop handling to the graph viewport and opens a preview", async () => {
-    const legendMessages = getDashboardFlowAxisLegendMessages("en");
-    const file = new File(["png"], "factory-import.png", { type: "image/png" });
-    const importValue = createFactoryImportValue();
-    const onFactoryImportReady =
-      vi.fn<(value: FactoryPngImportValue, file: File) => void>();
-    const readFactoryImportFile = vi
-      .fn<ReadFactoryImportFile>()
-      .mockResolvedValue({
-        ok: true,
-        value: importValue,
-      });
-
-    renderCurrentActivity({
-      onFactoryImportReady,
-      readFactoryImportFile,
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    const viewport = await screen.findByRole("region", {
-      name: "Work graph viewport",
-    });
-    const legendToggle = screen.getByRole("button", {
-      name: legendMessages.expandToggleLabel("graph legend"),
-    });
-
-    fireEvent.dragOver(legendToggle, createFileDropTransfer([file]));
-
-    expect(viewport.getAttribute("data-current-activity-drop-state")).toBe(
-      "idle",
-    );
-    expect(readFactoryImportFile).not.toHaveBeenCalled();
-
-    fireEvent.dragOver(viewport, createFileDropTransfer([file]));
-
-    expect(viewport.getAttribute("data-current-activity-drop-state")).toBe(
-      "drag-active",
-    );
-    expect(screen.getByText("Import factory PNG")).toBeTruthy();
-    expect(
-      screen.getByText(
-        "Drop a you-agent-factory PNG onto this graph to start import.",
-      ),
-    ).toBeTruthy();
-
-    fireEvent.drop(viewport, createFileDropTransfer([file]));
-
-    await waitFor(() => {
-      expect(readFactoryImportFile).toHaveBeenCalledWith(file);
-    });
-    await waitFor(() => {
-      expect(onFactoryImportReady).toHaveBeenCalledWith(importValue, file);
-    });
-    const previewDialog = await screen.findByRole("dialog", {
-      name: "Review factory import",
-    });
-
-    expect(previewDialog.textContent).toContain("Dropped Factory");
-    expect(previewDialog.textContent).toContain("factory-import.png");
-    expect(previewDialog.textContent).toContain(
-      "Review the dropped factory before confirming import.",
-    );
-    expect(
-      within(previewDialog)
-        .getByRole("img", { name: "Dropped Factory preview" })
-        .getAttribute("src"),
-    ).toBe("blob:factory-preview");
-    expect(viewport.getAttribute("data-current-activity-drop-state")).toBe(
-      "idle",
-    );
-
-    fireEvent.click(
-      within(previewDialog).getByRole("button", { name: "Cancel import" }),
-    );
-
-    await waitFor(() => {
-      expect(
-        screen.queryByRole("dialog", { name: "Review factory import" }),
-      ).toBeNull();
-    });
-    expect(importValue.revokePreviewImageSrc).toHaveBeenCalledTimes(1);
-    expect(
-      screen.getByRole("button", { name: "Select Review workstation" }),
-    ).toBeTruthy();
-  });
-
-  it("closes the factory import preview from the shared dialog close control", async () => {
-    const file = new File(["png"], "factory-import.png", { type: "image/png" });
-    const importValue = createFactoryImportValue();
-    const onFactoryImportReady =
-      vi.fn<(value: FactoryPngImportValue, file: File) => void>();
-    const readFactoryImportFile = vi
-      .fn<ReadFactoryImportFile>()
-      .mockResolvedValue({
-        ok: true,
-        value: importValue,
-      });
-
-    renderCurrentActivity({
-      onFactoryImportReady,
-      readFactoryImportFile,
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    const viewport = await screen.findByRole("region", {
-      name: "Work graph viewport",
-    });
-
-    fireEvent.drop(viewport, createFileDropTransfer([file]));
-
-    const previewDialog = await screen.findByRole("dialog", {
-      name: "Review factory import",
-    });
-    const closeButton = within(previewDialog).getByRole("button", {
-      name: "Close import preview",
-    });
-
-    fireEvent.click(closeButton);
-
-    await waitFor(() => {
-      expect(
-        screen.queryByRole("dialog", { name: "Review factory import" }),
-      ).toBeNull();
-    });
-    expect(importValue.revokePreviewImageSrc).toHaveBeenCalledTimes(1);
-    expect(onFactoryImportReady).toHaveBeenCalledWith(importValue, file);
-  });
-
-  it("does not render the import preview inside the graph card when a dashboard controller owns it", () => {
-    renderCurrentActivity({
-      importController: createImportController({
-        importPreviewState: {
-          file: new File(["png"], "factory-import.png", { type: "image/png" }),
-          status: "ready",
-          value: createFactoryImportValue(),
-        },
-      }),
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    expect(
-      screen.queryByRole("dialog", { name: "Review factory import" }),
-    ).toBeNull();
-    expect(
-      screen.getByRole("region", { name: "Work graph viewport" }),
-    ).toBeTruthy();
-  });
-
-  it("preserves the lean outer card shell while keeping the current activity region semantics", () => {
-    const legendMessages = getDashboardFlowAxisLegendMessages("en");
-
-    renderCurrentActivity({
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    const card = screen.getByLabelText("Current activity");
-    const legendToggle = screen.getByRole("button", {
-      name: legendMessages.expandToggleLabel("graph legend"),
-    });
-    const legend = legendToggle.closest(
-      "[data-dashboard-flow-axis-legend]",
-    ) as HTMLElement | null;
-    const viewport = screen.getByRole("region", {
-      name: "Work graph viewport",
-    });
-
-    expect(card?.className).toContain("relative");
-    expect(card?.className).toContain("flex");
-    expect(card?.className).toContain("h-full");
-    expect(card?.className).toContain("max-h-full");
-    expect(card?.className).toContain("min-h-0");
-    expect(card?.className).toContain("overflow-hidden");
-    expect((card as HTMLElement | null)?.style.height).toBe("100%");
-    expect((card as HTMLElement | null)?.style.maxHeight).toBe("100%");
-    expect((card as HTMLElement | null)?.style.overflow).toBe("hidden");
-    expect(card?.className).not.toMatch(PADDING_CLASS_PATTERN);
-    expect(
-      screen.getByRole("heading", { name: "Current activity" }),
-    ).toBeTruthy();
-    expect(screen.queryByText("Observe")).toBeNull();
-    expect(legend?.className).toContain("absolute");
-    expect(legend?.className).toContain("left-4");
-    expect(legend?.className).toContain("right-4");
-    expect(legend?.className).toContain("top-4");
-    expect(legend?.className).not.toMatch(PADDING_CLASS_PATTERN);
-    expect(viewport.className).toContain("max-h-full");
-    expect(viewport.className).toContain("overflow-hidden");
-    expect(viewport.style.height).toBe("100%");
-    expect(viewport.style.maxHeight).toBe("100%");
-    expect(viewport.style.overflow).toBe("hidden");
-    expect(viewport.className).not.toMatch(PADDING_CLASS_PATTERN);
-    expect(viewport.getAttribute("aria-describedby")).toMatch(
-      /^workflow-graph-heading-/,
-    );
-  });
-
-  it("renders a clear local alert when dropped PNG validation fails", async () => {
-    const file = new File(["png"], "invalid-factory.png", {
-      type: "image/png",
-    });
-    const onFactoryImportReady =
-      vi.fn<(value: FactoryPngImportValue, file: File) => void>();
-    const readFactoryImportFile = vi
-      .fn<ReadFactoryImportFile>()
-      .mockResolvedValue({
-        error: {
-          code: "PNG_METADATA_MISSING",
-          message:
-            "The selected PNG does not contain you-agent-factory factory metadata.",
-        },
-        ok: false,
-      });
-
-    renderCurrentActivity({
-      onFactoryImportReady,
-      readFactoryImportFile,
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    const viewport = await screen.findByRole("region", {
-      name: "Work graph viewport",
-    });
-
-    fireEvent.drop(viewport, createFileDropTransfer([file]));
-
-    const alert = await screen.findByRole("alert");
-
-    expect(alert.textContent).toContain("Factory import failed");
-    expect(alert.textContent).toContain("invalid-factory.png");
-    expect(alert.textContent).toContain(
-      "This PNG does not include the you-agent-factory factory metadata needed for import.",
-    );
-    expect(onFactoryImportReady).not.toHaveBeenCalled();
-    expect(
-      screen.queryByRole("dialog", { name: "Review factory import" }),
-    ).toBeNull();
-    expect(viewport.getAttribute("data-current-activity-drop-state")).toBe(
-      "error",
-    );
-    expect(
-      await screen.findByRole("button", { name: "Select Review workstation" }),
-    ).toBeTruthy();
-
-    fireEvent.click(screen.getByRole("button", { name: "Dismiss" }));
-
-    await waitFor(() => {
-      expect(screen.queryByRole("alert")).toBeNull();
-    });
-    expect(viewport.getAttribute("data-current-activity-drop-state")).toBe(
-      "idle",
-    );
-  });
-
-  it("activates the dropped factory, closes the preview, and requests an active-view refresh", async () => {
-    const file = new File(["png"], "factory-import.png", { type: "image/png" });
-    const importValue = createFactoryImportValue();
-    let resolveActivation: ((value: ImportFactoryValue) => void) | null = null;
-    const activateFactory = vi
-      .fn<(input: FactoryImportConfirmInput) => Promise<ImportFactoryValue>>()
-      .mockImplementation(
-        () =>
-          new Promise<ImportFactoryValue>((resolve) => {
-            resolveActivation = resolve;
-          }),
-      );
-    const onFactoryActivated = vi.fn<() => void>();
-    const readFactoryImportFile = vi
-      .fn<ReadFactoryImportFile>()
-      .mockResolvedValue({
-        ok: true,
-        value: importValue,
-      });
-
-    renderCurrentActivity({
-      activateFactory,
-      onFactoryActivated,
-      readFactoryImportFile,
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    const viewport = await screen.findByRole("region", {
-      name: "Work graph viewport",
-    });
-
-    fireEvent.drop(viewport, createFileDropTransfer([file]));
-
-    const previewDialog = await screen.findByRole("dialog", {
-      name: "Review factory import",
-    });
-
-    fireEvent.click(
-      within(previewDialog).getByRole("button", { name: "Confirm import" }),
-    );
-
-    await waitFor(() => {
-      expect(activateFactory).toHaveBeenCalledWith(
-        createFactoryImportConfirmInput(importValue, {
-          existingFactoryNames: ["dashboard-fixture", "Dropped Factory"],
-        }),
-      );
-    });
-    const activateButton = within(previewDialog).getByRole<HTMLButtonElement>(
-      "button",
-      {
-        name: "Activating factory...",
-      },
-    );
-    const cancelButton = within(previewDialog).getByRole<HTMLButtonElement>(
-      "button",
-      {
-        name: "Cancel import",
-      },
-    );
-    const closeButton = within(previewDialog).getByRole<HTMLButtonElement>(
-      "button",
-      {
-        name: "Close import preview",
-      },
-    );
-
-    expect(activateButton.getAttribute("aria-busy")).toBe("true");
-    expect(activateButton.disabled).toBe(true);
-    expect(cancelButton.disabled).toBe(true);
-    expect(closeButton.disabled).toBe(true);
-
-    resolveActivation?.(importValue.factory);
-
-    await waitFor(() => {
-      expect(onFactoryActivated).toHaveBeenCalledTimes(1);
-    });
-    await waitFor(() => {
-      expect(
-        screen.queryByRole("dialog", { name: "Review factory import" }),
-      ).toBeNull();
-    });
-    expect(importValue.revokePreviewImageSrc).toHaveBeenCalledTimes(1);
-  });
-
-  it("shows a distinct duplicate-name activation error without changing the current view", async () => {
-    const file = new File(["png"], "factory-import.png", { type: "image/png" });
-    const importValue = createFactoryImportValue();
-    const activateFactory = vi
-      .fn<(input: FactoryImportConfirmInput) => Promise<ImportFactoryValue>>()
-      .mockRejectedValue(
-        new SessionFactoryAPIError("Named factory already exists.", {
-          code: "FACTORY_ALREADY_EXISTS",
-          status: 409,
-        }),
-      );
-    const onFactoryActivated = vi.fn<() => void>();
-    const readFactoryImportFile = vi
-      .fn<ReadFactoryImportFile>()
-      .mockResolvedValue({
-        ok: true,
-        value: importValue,
-      });
-
-    renderCurrentActivity({
-      activateFactory,
-      onFactoryActivated,
-      readFactoryImportFile,
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    const viewport = await screen.findByRole("region", {
-      name: "Work graph viewport",
-    });
-
-    fireEvent.drop(viewport, createFileDropTransfer([file]));
-
-    const previewDialog = await screen.findByRole("dialog", {
-      name: "Review factory import",
-    });
-
-    fireEvent.click(
-      within(previewDialog).getByRole("button", { name: "Confirm import" }),
-    );
-
-    const alert = await within(previewDialog).findByRole("alert");
-
-    expect(activateFactory).toHaveBeenCalledWith(
-      createFactoryImportConfirmInput(importValue, {
-        existingFactoryNames: ["dashboard-fixture", "Dropped Factory"],
-      }),
-    );
-    expect(alert.textContent).toContain("Activation failed");
-    expect(alert.textContent).toContain(
-      "A factory with this name already exists.",
-    );
-    expect(onFactoryActivated).not.toHaveBeenCalled();
-    expect(
-      screen.getByRole("dialog", { name: "Review factory import" }),
-    ).toBeTruthy();
-    expect(importValue.revokePreviewImageSrc).not.toHaveBeenCalled();
-  });
-
-  it("shows a distinct non-idle activation error without changing the current view", async () => {
-    const file = new File(["png"], "factory-import.png", { type: "image/png" });
-    const importValue = createFactoryImportValue();
-    const activateFactory = vi
-      .fn<(input: FactoryImportConfirmInput) => Promise<ImportFactoryValue>>()
-      .mockRejectedValue(
-        new SessionFactoryAPIError(
-          "Current factory runtime must be idle before activation.",
-          {
-            code: "FACTORY_NOT_IDLE",
-            status: 409,
-          },
-        ),
-      );
-    const onFactoryActivated = vi.fn<() => void>();
-    const readFactoryImportFile = vi
-      .fn<ReadFactoryImportFile>()
-      .mockResolvedValue({
-        ok: true,
-        value: importValue,
-      });
-
-    renderCurrentActivity({
-      activateFactory,
-      onFactoryActivated,
-      readFactoryImportFile,
-      snapshot: semanticWorkflowDashboardSnapshot,
-    });
-
-    const viewport = await screen.findByRole("region", {
-      name: "Work graph viewport",
-    });
-
-    fireEvent.drop(viewport, createFileDropTransfer([file]));
-
-    const previewDialog = await screen.findByRole("dialog", {
-      name: "Review factory import",
-    });
-
-    fireEvent.click(
-      within(previewDialog).getByRole("button", { name: "Confirm import" }),
-    );
-
-    const alert = await within(previewDialog).findByRole("alert");
-
-    expect(activateFactory).toHaveBeenCalledWith(
-      createFactoryImportConfirmInput(importValue, {
-        existingFactoryNames: ["dashboard-fixture", "Dropped Factory"],
-      }),
-    );
-    expect(alert.textContent).toContain("Activation failed");
-    expect(alert.textContent).toContain(
-      "The current factory runtime is still active.",
-    );
-    expect(onFactoryActivated).not.toHaveBeenCalled();
-    expect(
-      screen.getByRole("dialog", { name: "Review factory import" }),
-    ).toBeTruthy();
-    expect(importValue.revokePreviewImageSrc).not.toHaveBeenCalled();
-  });
-});
 
 describe("ReactFlowCurrentActivityCard graph semantics", () => {
   registerCurrentActivityCardTestLifecycle();
@@ -3102,8 +1092,7 @@ describe("ReactFlowCurrentActivityCard graph semantics", () => {
     expect(screen.getByLabelText("agent-slot")).toBeTruthy();
 
     cleanup();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
+    reinstallDashboardBrowserTestShims();
 
     const activeSnapshot = resourceOccupancySnapshotForTick(3);
 
@@ -3122,8 +1111,7 @@ describe("ReactFlowCurrentActivityCard graph semantics", () => {
     expect(screen.queryByLabelText("2 resource tokens")).toBeNull();
 
     cleanup();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
+    reinstallDashboardBrowserTestShims();
 
     const returnedSnapshot = resourceOccupancySnapshotForTick(4);
 
@@ -3203,8 +1191,7 @@ describe("ReactFlowCurrentActivityCard graph semantics", () => {
     );
 
     cleanup();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
+    reinstallDashboardBrowserTestShims();
 
     renderCurrentActivity({
       snapshot: semanticWorkflowDashboardSnapshot,
